refactor of the challenge response model
there are multiple problems with the previous model, it sends each block header one by one, we can get smarter and do an spv proof for each block which is smaller and then compact calls into proofs and send them, at the end doing a difficulty algorithm check on the entire superblock. By doing one call or a few calls it saves gas aswell as time for a challenge, now the submitter has 10 minutes to submit all of his proofs so it sets an upperbound for a challenge period whereas before it was unbounded based on number of block hashes.