fix web3j export by converting lastBlockInfo to mapping of blockInfo to byte32
for some reason web3j gives a tuple error on export, so I had to convert:

"BlockInfo lastBlockInfo;"

to

"mapping (bytes32 => BlockInfo) blocksInfo;"

There will only ever be one entry to blocksInfo (it will simply be the block tied to the superblock (the last block), it used to be the array of block headers before we created a new challenge model, but we didn't need the array of block headers (just need the last header) so I removed the mapping and replaced with just one BlockInfo but that seem to break it. The mapping works however we only will need one entry (session.blocksInfo[session.superblockHash]).