fix lastBlock hijack attack
a bad defender can craft a superblock such that the lastblock+prevblock of last are valid but remaining 58 hashes are not linked to the lastblock and thus we need to have challengers watch for links that are broken in the chain of blocks.

Therefor we need to store all 60 hashes of the superblock in the session and challengers can verify if they see a difference between a common block and the prev block which is different. We therefor need to pass up to 1 more block header in those cases where lastblock is valid but the link of hashes is broken between 0 and hash 57.