remove unneeded event emissions
tokenunfreeze on superblocksubmitter unless it helps with etherscan somehow is kind of misleading as the unfreezing is really happening on the owner and you are transferring some funds to the superblock submitter as a fee.

the other verify/relay transaction are only set to 0 for txhash in case of errors, we don't really need to check for a non-zero value for a successful case because we can relay on the fact that we "dont" see these events, because if we do its likely an error, also token unfreeze and transfer can be used to detect that a token has been unfrozen.