work on #25
In order to prevent #25 we need to be able to prove the first block->prev hash matches the prev superblock last block hash. This will enforce linking of headers between superblocks from last to first of next superblock. -1 is now used as a constant for interim index to mean that no interim blocks are requested, and 0 will mean check for 0th index. We previously used 0 to denote that no interim blocks are needed. We were missing the check at 0th index.