fix adjustmentpow test, all passing gas at 5m
we only need to check for auxiliary work on the last block. We need to ensure the PoW is valid for all blocks and that they hash to each other but we only need to verify the coinbase merkle and the parent merkle of auxiliary work on the last block since it is assumed the blocks being connected that the aux work on previous headers is also correct (otherwise they would not hash to the last block that is valid).